+++ aldrin/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Add `Handle::wait_for_object()`, `Handle::wait_for_any_object()` and
+  `Handle::wait_for_specific_object()`.
+
 ## [0.3.0] - 2024-01-18
 
 ### Added
@@ -48,6 +55,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
+[Unreleased]: https://github.com/dennis-hamester/aldrin/compare/aldrin-0.3.0...HEAD
 [0.3.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-0.3.0
 [0.2.1]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-0.2.1
 [0.2.0]: https://github.com/dennis-hamester/aldrin/releases/tag/aldrin-0.2.0